+++ Assets/Demo/SimpleDemo/SimpleController_UsingActions_InAsset.cs
@@ -12,7 +12,7 @@ public class SimpleController_UsingActions_InAsset : MonoBehaviour
     public float jumpForce = 2.0f;
     public GameObject projectile;
 
-    public SimpleControls controls;
+    private SimpleControls controls;
 
     private Vector2 m_Move;
     private Vector2 m_Look;
@@ -22,7 +22,7 @@ public class SimpleController_UsingActions_InAsset : MonoBehaviour
     private Rigidbody m_Rigidbody;
 
     private void Start()
-    {
+    {		
         m_Rigidbody = GetComponent<Rigidbody>();
 
         ////FIXME: Solve this properly. ATM, if we have both fixed and dynamic updates enabled, then
@@ -40,6 +40,7 @@ public class SimpleController_UsingActions_InAsset : MonoBehaviour
 
     public void Awake()
     {
+		controls = new SimpleControls();
         controls.gameplay.move.performed += ctx => m_Move = ctx.ReadValue<Vector2>();
         controls.gameplay.look.performed += ctx => m_Look = ctx.ReadValue<Vector2>();
         controls.gameplay.move.cancelled += ctx => m_Move = Vector2.zero;