+++ Packages/com.unity.inputsystem/InputSystem/Editor/AssetImporter/InputActionImporterEditor.cs
@@ -37,15 +37,15 @@ namespace UnityEngine.Experimental.Input.Editor
                 var wrapperCodePathProperty = serializedObject.FindProperty("m_WrapperCodePath");
                 var wrapperClassNameProperty = serializedObject.FindProperty("m_WrapperClassName");
                 var wrapperCodeNamespaceProperty = serializedObject.FindProperty("m_WrapperCodeNamespace");
-                var generateActionEventsProperty = serializedObject.FindProperty("m_GenerateActionEvents");
-                var generateInterfacesProperty = serializedObject.FindProperty("m_GenerateInterfaces");
 
                 EditorGUILayout.BeginHorizontal();
-                EditorGUILayout.PropertyField(wrapperCodePathProperty, m_WrapperCodePathLabel);
-                if (GUILayout.Button("...", EditorStyles.miniButton, GUILayout.MaxWidth(20)))
+                var assetPath = AssetDatabase.GetAssetPath(GetAsset());
+                var defaultFileName = Path.ChangeExtension(assetPath, ".cs");
+
+                wrapperCodePathProperty.PropertyFieldWithDefaultText(m_WrapperCodePathLabel, defaultFileName);
+
+                if (GUILayout.Button("…", EditorStyles.miniButton, GUILayout.MaxWidth(20)))
                 {
-                    var assetPath = AssetDatabase.GetAssetPath(GetAsset());
-                    var defaultFileName = Path.ChangeExtension(assetPath, ".cs");
                     var fileName = EditorUtility.SaveFilePanel("Location for generated C# file",
                         Path.GetDirectoryName(defaultFileName),
                         Path.GetFileName(defaultFileName), "cs");
@@ -59,16 +59,15 @@ namespace UnityEngine.Experimental.Input.Editor
                 }
                 EditorGUILayout.EndHorizontal();
 
-                EditorGUILayout.PropertyField(wrapperClassNameProperty, m_WrapperClassNameLabel);
+                wrapperClassNameProperty.PropertyFieldWithDefaultText(m_WrapperClassNameLabel, CSharpCodeHelpers.MakeTypeName(GetAsset().name));
+
                 if (!CSharpCodeHelpers.IsEmptyOrProperIdentifier(wrapperClassNameProperty.stringValue))
                     EditorGUILayout.HelpBox("Must be a valid C# identifier", MessageType.Error);
 
-                EditorGUILayout.PropertyField(wrapperCodeNamespaceProperty, m_WrapperCodeNamespaceLabel);
+                wrapperCodeNamespaceProperty.PropertyFieldWithDefaultText(m_WrapperCodeNamespaceLabel, "<Global namespace>");
+
                 if (!CSharpCodeHelpers.IsEmptyOrProperNamespaceName(wrapperCodeNamespaceProperty.stringValue))
                     EditorGUILayout.HelpBox("Must be a valid C# namespace name", MessageType.Error);
-
-                EditorGUILayout.PropertyField(generateActionEventsProperty, m_GenerateActionEventsLabel);
-                EditorGUILayout.PropertyField(generateInterfacesProperty);
             }
 
             #if UNITY_2019_2_OR_NEWER
@@ -89,7 +88,6 @@ namespace UnityEngine.Experimental.Input.Editor
         }
 
         private readonly GUIContent m_GenerateWrapperCodeLabel = EditorGUIUtility.TrTextContent("Generate C# Class");
-        private readonly GUIContent m_GenerateActionEventsLabel = EditorGUIUtility.TrTextContent("Generate Events");
         private readonly GUIContent m_WrapperCodePathLabel = EditorGUIUtility.TrTextContent("C# Class File");
         private readonly GUIContent m_WrapperClassNameLabel = EditorGUIUtility.TrTextContent("C# Class Name");
         private GUIContent m_WrapperCodeNamespaceLabel = EditorGUIUtility.TrTextContent("C# Class Namespace");