+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - `Joystick.axes` and `Joystick.buttons` have been removed.
+- Generated wrapper code for Input Action Assets are now self-contained, generating all the data from code and not needing a reference to the asset; `InputActionAssetReference` has been removed.
+- The option to generate interfaces on wrappers has been removed, instead we always do this now.
+- The option to generate events on wrappers has been removed, we felt that this no longer made sense.
+- Will now show default values in Input Action inspector if no custom values for file path, class name or namespace have been provided.
 - `InputSettings.runInBackground` has been removed. This should now be supported or not on a per-device level. Most devices never supported it in the first place, so a global setting did not seem to be useful.
 
 ### Fixed